Renaming ProvenanceMarshaller to ProvenanceSerialization.
