+++ olcut-config-json/src/test/java/com/oracle/labs/mlrg/olcut/config/json/JsonProvenanceSerializationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Oracle and/or its affiliates.
+ * Copyright (c) 2004-2021, Oracle and/or its affiliates.
  *
  * Licensed under the 2-clause BSD license.
  *
@@ -26,13 +26,17 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.oracle.labs.mlrg.olcut.provenance;
+package com.oracle.labs.mlrg.olcut.config.json;
 
 import com.oracle.labs.mlrg.olcut.config.ConfigurationManager;
+import com.oracle.labs.mlrg.olcut.provenance.ExampleProvenancableConfigurable;
+import com.oracle.labs.mlrg.olcut.provenance.ListProvenance;
+import com.oracle.labs.mlrg.olcut.provenance.ObjectProvenance;
+import com.oracle.labs.mlrg.olcut.provenance.Provenance;
 import com.oracle.labs.mlrg.olcut.provenance.ProvenanceConversionTest.SimpleObjectProvenance;
+import com.oracle.labs.mlrg.olcut.provenance.ProvenanceUtil;
 import com.oracle.labs.mlrg.olcut.provenance.io.ObjectMarshalledProvenance;
 import com.oracle.labs.mlrg.olcut.provenance.io.ProvenanceSerializationException;
-import com.oracle.labs.mlrg.olcut.provenance.io.XMLProvenanceMarshaller;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -47,12 +51,12 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 /**
  *
  */
-public class XmlProvenanceConversionTest {
-    private XMLProvenanceMarshaller marshaller;
+public class JsonProvenanceSerializationTest {
+    private final JsonProvenanceSerialization marshaller = new JsonProvenanceSerialization(true);
 
     @BeforeEach
     public void setUp() throws IOException {
-        marshaller = new XMLProvenanceMarshaller(true);
+        ConfigurationManager.addFileFormatFactory(new JsonConfigFactory());
     }
 
     @Test
@@ -66,11 +70,11 @@ public class XmlProvenanceConversionTest {
         List<ObjectMarshalledProvenance> marshalledProvenances = ProvenanceUtil.marshalProvenance(provenance);
         assertEquals(8,marshalledProvenances.size());
 
-        String xmlResult = marshaller.serializeToString(marshalledProvenances);
+        String jsonResult = marshaller.serializeToString(marshalledProvenances);
 
-        List<ObjectMarshalledProvenance> xmlProvenances = marshaller.deserializeFromString(xmlResult);
+        List<ObjectMarshalledProvenance> jsonProvenances = marshaller.deserializeFromString(jsonResult);
 
-        ObjectProvenance unmarshalledProvenance = ProvenanceUtil.unmarshalProvenance(xmlProvenances);
+        ObjectProvenance unmarshalledProvenance = ProvenanceUtil.unmarshalProvenance(jsonProvenances);
 
         assertEquals(provenance,unmarshalledProvenance);
     }
@@ -87,11 +91,11 @@ public class XmlProvenanceConversionTest {
 
         assertEquals(1,marshalledProvenance.size());
 
-        String xmlResult = marshaller.serializeToString(marshalledProvenance);
+        String jsonResult = marshaller.serializeToString(marshalledProvenance);
 
-        List<ObjectMarshalledProvenance> xmlProvenances = marshaller.deserializeFromString(xmlResult);
+        List<ObjectMarshalledProvenance> jsonProvenances = marshaller.deserializeFromString(jsonResult);
 
-        ObjectProvenance unmarshalledProvenance = ProvenanceUtil.unmarshalProvenance(xmlProvenances);
+        ObjectProvenance unmarshalledProvenance = ProvenanceUtil.unmarshalProvenance(jsonProvenances);
 
         assertEquals(objProv,unmarshalledProvenance);
     }
